// src/core/interfaces/movie-credits.interface.ts
import { MovieDetail } from "./movie-detail.interface";

export interface Cast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface Crew {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  credit_id: string;
  department: string;
  job: string;
}

export interface CreditResponse {
  cast: Cast[];
  crew: Crew[];
}

export interface DetailInfoProps {
  item: MovieDetail;
  credit: CreditResponse;
}